requirements.txt
flask
flask-Cors
requests
python-dotenv
flake8
entsoe-py
pandas
pulp
matplotlib